Unify description and types mechanism to work on Triple type instead of NetworkTriple
