+++ apps/web/modules/entity/entity.test.ts
@@ -1,8 +1,7 @@
 import { describe, expect, it } from 'vitest';
 import { SYSTEM_IDS } from '../constants';
-import { NetworkEntity } from '../services/network';
 import { Triple } from '../types';
-import { networkStringDescriptionValue, description } from './entity';
+import { description } from './entity';
 
 const triplesWithSystemDescriptionAttribute: Triple[] = [
   {
@@ -52,44 +51,6 @@ const triplesWithNonSystemDescriptionAttributeAndValueIsEntity: Triple[] = [
   },
 ];
 
-// A triple may be referencing a Description different from the expected
-// system one.
-const networkTriplesWithDescriptionName: NetworkEntity['entityOf'] = [
-  {
-    attribute: {
-      id: 'attributeId',
-      name: 'Description',
-    },
-    stringValue: 'banana',
-    valueType: 'STRING',
-  },
-];
-
-const networkTriplesWithDescriptionId: NetworkEntity['entityOf'] = [
-  {
-    attribute: {
-      id: SYSTEM_IDS.DESCRIPTION_SCALAR,
-      name: '',
-    },
-    stringValue: 'banana',
-    valueType: 'STRING',
-  },
-];
-
-const networkTriplesWithDescriptionAsEntityValue: NetworkEntity['entityOf'] = [
-  {
-    attribute: {
-      id: SYSTEM_IDS.DESCRIPTION_SCALAR,
-      name: '',
-    },
-    entityValue: {
-      id: 'entityId',
-      name: 'banana',
-    },
-    valueType: 'ENTITY',
-  },
-];
-
 /**
  * We assume that the Description triple's attribute for an Entity will match the expected
  * system Description attribute ID at SYSTEM_IDS.DESCRIPTION_SCALAR. However, anybody can
@@ -112,17 +73,5 @@ describe('Entity description helpers', () => {
 
   it('Parses description from triples where description is not the expected system Description and value is a reference to another Entity', () => {
     expect(description(triplesWithNonSystemDescriptionAttributeAndValueIsEntity)).toBe(null);
-  });
-
-  it('Parses description from network GeoEntity where description attribute is not the expected system Description', () => {
-    expect(networkStringDescriptionValue(networkTriplesWithDescriptionName)).toBe('banana');
-  });
-
-  it('Parses description from network GeoEntity where description attribute is the expected system Description', () => {
-    expect(networkStringDescriptionValue(networkTriplesWithDescriptionId)).toBe('banana');
-  });
-
-  it('Returns undefined from network GeoEntity where description value is a reference to another Entity', () => {
-    expect(networkStringDescriptionValue(networkTriplesWithDescriptionAsEntityValue)).toBe(null);
   });
 });
